+++ R/emfrail_aux.R
@@ -97,7 +97,7 @@ laplace_transform <- function(x, .distribution) {
 #'
 #' fr_var <- c(0.001, 0.01, 0.1, 0.2, 0.3, 0.4, 0.5, 0.6, 0.7, 0.8, 0.9)
 #' profloglik <- emfrail_pll(rats,
-#'                           Surv(rep(time, status) ~ rx + sex + cluster(litter),
+#'                           Surv(time, status) ~ rx + sex + cluster(litter),
 #'                           .values = 1/fr_var)
 #' plot(fr_var, profloglik, xlab = "frailty variance", ylab = "profile log-likelihood")
 #'
@@ -124,6 +124,7 @@ emfrail_pll <- function(.data, .formula,
 
 
 }
+
 
 
 # this one gives the baseline cumulative hazard at all the time points;