+++ NEWS.md
@@ -1,3 +1,13 @@
+# frailtoys 0.4.7
+Added a `newdata` option for the `predict` method and for the `plot` methods. 
+This can be used instead of `lp`, and basically calculates the corresponding linear predictor for the 
+given covariate values.
+
+bugfixes
+
+# frailtoys 0.4.6
+There is an option now to calculate the unadjusted SE or no SE at all
+
 # frailtoys 0.4.5
 
 * There are now plot methods available! Check out `?plot_emfrail`\